+++ packages/applications/bootstrap/src/setupTâche.ts
@@ -1,6 +1,12 @@
 import { mediator } from 'mediateur';
 
-import { registerTâcheCommand, registerTâcheQuery, TâcheSaga } from '@potentiel-domain/tache';
+import {
+  TâcheAbandonSaga,
+  TâcheGarantiesFinancièresSaga,
+  TâcheRaccordementSaga,
+  registerTâcheCommand,
+  registerTâcheQuery,
+} from '@potentiel-domain/tache';
 import { Event, loadAggregate, subscribe } from '@potentiel-infrastructure/pg-event-sourcing';
 import { TâcheProjector } from '@potentiel-applications/projectors';
 import { récupérerIdentifiantsProjetParEmailPorteurAdapter } from '@potentiel-infrastructure/domain-adapters';
@@ -17,63 +23,22 @@ export const setupTâche = async () => {
     list: listProjection,
   });
 
-  TâcheSaga.register();
+  const unsubscribeTâcheProjector = await registerTâcheProjector();
+
+  const unsubscribeTâcheAbandonSaga = await registerTâcheAbandonSaga();
+  const unsubscribeTâcheRaccordementSaga = await registerTâcheRaccordementSaga();
+  const unsubscribeTâcheGarantiesFinancièresSaga = await registerTâcheGarantiesFinancières();
+
+  return async () => {
+    await unsubscribeTâcheAbandonSaga();
+    await unsubscribeTâcheRaccordementSaga();
+    await unsubscribeTâcheGarantiesFinancièresSaga();
+    await unsubscribeTâcheProjector();
+  };
+};
+
+const registerTâcheProjector = async () => {
   TâcheProjector.register();
-
-  const unsubscribeTâcheAbandonSaga = await subscribe<TâcheSaga.AbandonSubscriptionEvent & Event>({
-    name: 'tache-saga',
-    streamCategory: 'abandon',
-    eventType: [
-      'AbandonAnnulé-V1',
-      'AbandonConfirmé-V1',
-      'AbandonRejeté-V1',
-      'ConfirmationAbandonDemandée-V1',
-      'PreuveRecandidatureDemandée-V1',
-      'PreuveRecandidatureTransmise-V1',
-    ],
-    eventHandler: async (event) => {
-      await mediator.publish<TâcheSaga.Execute>({
-        type: 'System.Saga.Tâche',
-        data: event,
-      });
-    },
-  });
-
-  const unsubscribeTâcheRaccordementSaga = await subscribe<
-    TâcheSaga.RaccordementSubscriptionEvent & Event
-  >({
-    name: 'tache-saga',
-    streamCategory: 'raccordement',
-    eventType: [
-      'RéférenceDossierRacordementModifiée-V1',
-      'GestionnaireRéseauRaccordementModifié-V1',
-      'GestionnaireRéseauInconnuAttribué-V1',
-    ],
-    eventHandler: async (event) => {
-      await mediator.publish<TâcheSaga.Execute>({
-        type: 'System.Saga.Tâche',
-        data: event,
-      });
-    },
-  });
-
-  const unsubscribeTâcheGarantiesFinancièresSaga = await subscribe<
-    TâcheSaga.GarantiesFinancièresSubscriptionEvent & Event
-  >({
-    name: 'tache-saga',
-    streamCategory: 'garanties-financieres',
-    eventType: [
-      'GarantiesFinancièresDemandées-V1',
-      'DépôtGarantiesFinancièresSoumis-V1',
-      'GarantiesFinancièresEnregistrées-V1',
-    ],
-    eventHandler: async (event) => {
-      await mediator.publish<TâcheSaga.Execute>({
-        type: 'System.Saga.Tâche',
-        data: event,
-      });
-    },
-  });
 
   const unsubscribeTâcheProjector = await subscribe<TâcheProjector.SubscriptionEvent>({
     name: 'projector',
@@ -92,11 +57,72 @@ export const setupTâche = async () => {
     },
     streamCategory: 'tâche',
   });
-
-  return async () => {
-    await unsubscribeTâcheAbandonSaga();
-    await unsubscribeTâcheRaccordementSaga();
-    await unsubscribeTâcheGarantiesFinancièresSaga();
-    await unsubscribeTâcheProjector();
-  };
+  return unsubscribeTâcheProjector;
 };
+
+const registerTâcheGarantiesFinancières = async () => {
+  TâcheGarantiesFinancièresSaga.register();
+  const unsubscribeTâcheGarantiesFinancièresSaga = await subscribe<
+    TâcheGarantiesFinancièresSaga.SubscriptionEvent & Event
+  >({
+    name: 'tache-saga',
+    streamCategory: 'garanties-financieres',
+    eventType: [
+      'GarantiesFinancièresDemandées-V1',
+      'DépôtGarantiesFinancièresSoumis-V1',
+      'GarantiesFinancièresEnregistrées-V1',
+    ],
+    eventHandler: async (event) => {
+      await mediator.publish<TâcheGarantiesFinancièresSaga.Execute>({
+        type: 'System.Saga.TâcheGarantiesFinancières',
+        data: event,
+      });
+    },
+  });
+  return unsubscribeTâcheGarantiesFinancièresSaga;
+};
+
+const registerTâcheRaccordementSaga = async () => {
+  TâcheRaccordementSaga.register();
+  const unsubscribeTâcheRaccordementSaga = await subscribe<
+    TâcheRaccordementSaga.SubscriptionEvent & Event
+  >({
+    name: 'tache-saga',
+    streamCategory: 'raccordement',
+    eventType: [
+      'RéférenceDossierRacordementModifiée-V1',
+      'GestionnaireRéseauRaccordementModifié-V1',
+      'GestionnaireRéseauInconnuAttribué-V1',
+    ],
+    eventHandler: async (event) => {
+      await mediator.publish<TâcheRaccordementSaga.Execute>({
+        type: 'System.Saga.TâcheRaccordement',
+        data: event,
+      });
+    },
+  });
+  return unsubscribeTâcheRaccordementSaga;
+};
+
+async function registerTâcheAbandonSaga() {
+  TâcheAbandonSaga.register();
+  const unsubscribeTâcheAbandonSaga = await subscribe<TâcheAbandonSaga.SubscriptionEvent & Event>({
+    name: 'tache-saga',
+    streamCategory: 'abandon',
+    eventType: [
+      'AbandonAnnulé-V1',
+      'AbandonConfirmé-V1',
+      'AbandonRejeté-V1',
+      'ConfirmationAbandonDemandée-V1',
+      'PreuveRecandidatureDemandée-V1',
+      'PreuveRecandidatureTransmise-V1',
+    ],
+    eventHandler: async (event) => {
+      await mediator.publish<TâcheAbandonSaga.Execute>({
+        type: 'System.Saga.TâcheAbandon',
+        data: event,
+      });
+    },
+  });
+  return unsubscribeTâcheAbandonSaga;
+}
